Click packet now contains a position; Made all coordinates in dc.rs f32; Added a 'poof' for clicks outside the circle

+++ client/src/dc.rs
@@ -16,15 +16,15 @@ extern {
 	fn dc_fill_color_raw(r: i32, g: i32, b: i32, a: f32);
 	fn dc_stroke_color_raw(r: i32, g: i32, b: i32, a: f32);
 
-	pub fn dc_fill_rect(x: i32, y: i32, w: i32, h: i32);
-	pub fn dc_fill_circle(x: i32, y: i32, r: f32);
-	pub fn dc_draw_circle(x: i32, y: i32, r: f32);
+	pub fn dc_fill_rect(x: f32, y: f32, w: f32, h: f32);
+	pub fn dc_fill_circle(x: f32, y: f32, r: f32);
+	pub fn dc_draw_circle(x: f32, y: f32, r: f32);
 
-	fn dc_fill_text_raw(t: *const c_char, x: i32, y: i32);
+	fn dc_fill_text_raw(t: *const c_char, x: f32, y: f32);
 	fn dc_set_font_raw(f: *const c_char);
 
-	pub fn get_canvas_height() -> i32;
-	pub fn get_canvas_width() -> i32;
+	pub fn get_canvas_height() -> f32;
+	pub fn get_canvas_width() -> f32;
 }
 
 pub unsafe fn dc_fill_color(col: Color) {
@@ -37,7 +37,7 @@ pub unsafe fn dc_stroke_color(col: Color) {
 	dc_stroke_color_raw(r as i32, g as i32, b as i32, col.a);
 }
 
-pub unsafe fn dc_fill_text(t: &str, x: i32, y: i32) {
+pub unsafe fn dc_fill_text(t: &str, x: f32, y: f32) {
 	let cstr = CString::new(t).unwrap();
 	dc_fill_text_raw(cstr.as_bytes_with_nul().as_ptr(), x, y);
 }
@@ -47,7 +47,7 @@ pub unsafe fn dc_set_font(t: &str) {
 	dc_set_font_raw(cstr.as_bytes_with_nul().as_ptr());
 }
 
-pub unsafe fn get_canvas_size() -> (i32,i32) {
+pub unsafe fn get_canvas_size() -> (f32,f32) {
 	(get_canvas_width(), get_canvas_height())
 }
